move: core:ui package -> core:ui module
