+++ core/ui/src/main/res/drawable/ic_bookmark_border_24.xml
