+++ core/ui/src/main/res/drawable/ic_bookmark_filled_24.xml
