+++ core/ui/src/main/res/drawable/ic_clear_24.xml
