+++ core/ui/src/main/res/drawable/ic_search_24.xml
